requirements.txt
pdfplumber~=0.7.4
camelot-py[cv]
pandas~=1.4.3